requirements.txt
shopifyapi~=12.7.0
flask~=3.1.0
python-decouple==3.8
Flask-SQLAlchemy==3.0.3
requests~=2.32.3
Faker==18.9.0